+++ libs/cardano-ledger-binary/test/Main.hs
@@ -4,6 +4,7 @@ import System.IO (BufferMode (LineBuffering), hSetBuffering, hSetEncoding, stdou
 import qualified Test.Cardano.Ledger.Binary.Failure as Failure
 import qualified Test.Cardano.Ledger.Binary.PlainSpec as PlainSpec
 import qualified Test.Cardano.Ledger.Binary.RoundTripSpec as RoundTripSpec
+import qualified Test.Cardano.Ledger.Binary.Success as Success
 import qualified Test.Cardano.Ledger.Binary.Vintage.Coders as Vintage.Coders
 import qualified Test.Cardano.Ledger.Binary.Vintage.Drop as Vintage.Drop
 import qualified Test.Cardano.Ledger.Binary.Vintage.Failure as Vintage.Failure
@@ -25,6 +26,7 @@ spec = do
   describe "Versioned" $ do
     RoundTripSpec.spec
     Failure.spec
+    Success.spec
 
 main :: IO ()
 main = do